Bug fixes and additional code
I was able to fix the bug that would exit the rematch loop. I was passing the address of the check flag among functions so it was messing with the same variable from the original function where it originated. Coded this so it would run the loop regardless but te value is still addressed correctly when choosing the proper response. I added some code to check for error bound on the rematch loop and helpful comments.